+++ contracts/shard/ShardVaultProxy.sol
@@ -20,13 +20,13 @@ contract ShardVaultProxy is Proxy {
         address jpegdVault,
         address jpegdVaultHelper,
         uint256 shardValue,
-        uint256 maxSupply,
-        uint256 id,
-        uint256 saleFeeBP,
-        uint256 acquisitionFeeBP,
-        uint256 yieldFeeBP,
-        uint256 bufferBP,
-        uint256 deviationBP
+        uint16 maxSupply,
+        uint16 id,
+        uint16 saleFeeBP,
+        uint16 acquisitionFeeBP,
+        uint16 yieldFeeBP,
+        uint16 bufferBP,
+        uint16 deviationBP
     ) {
         SHARD_VAULT_DIAMOND = shardVaultDiamond;
